+++ packages/stream_chat/lib/src/client/channel.dart
@@ -423,6 +423,14 @@ class Channel {
     return state!.channelStateStream.map((cs) => cs.channel?.messageCount);
   }
 
+  /// List of filter tags applied to this channel.
+  ///
+  /// Generally used for filtering channels while querying.
+  List<String>? get filterTags {
+    _checkInitialized();
+    return state!._channelState.channel?.filterTags;
+  }
+
   /// Channel id.
   String? get id => state?._channelState.channel?.id ?? _id;
 
@@ -1714,6 +1722,7 @@ class Channel {
     List<String> memberIds, {
     Message? message,
     bool hideHistory = false,
+    DateTime? hideHistoryBefore,
   }) async {
     _checkInitialized();
     return _client.addChannelMembers(
@@ -1722,6 +1731,7 @@ class Channel {
       memberIds,
       message: message,
       hideHistory: hideHistory,
+      hideHistoryBefore: hideHistoryBefore,
     );
   }
 
@@ -1780,10 +1790,9 @@ class Channel {
     return _client.markChannelRead(id!, type, messageId: messageId);
   }
 
-  /// Mark message as unread.
+  /// Marks the channel as unread by a given [messageId].
   ///
-  /// You have to provide a [messageId] from which you want the channel
-  /// to be marked as unread.
+  /// All messages from the provided message onwards will be marked as unread.
   Future<EmptyResponse> markUnread(String messageId) async {
     _checkInitialized();
 
@@ -1795,6 +1804,22 @@ class Channel {
     }
 
     return _client.markChannelUnread(id!, type, messageId);
+  }
+
+  /// Marks the channel as unread by a given [timestamp].
+  ///
+  /// All messages after the provided timestamp will be marked as unread.
+  Future<EmptyResponse> markUnreadByTimestamp(DateTime timestamp) async {
+    _checkInitialized();
+
+    if (!canUseReadReceipts) {
+      throw const StreamChatError(
+        'Cannot mark as unread: Channel does not support read events. '
+        'Enable read_events in your channel type configuration.',
+      );
+    }
+
+    return _client.markChannelUnreadByTimestamp(id!, type, timestamp);
   }
 
   /// Mark the thread with [threadId] in the channel as read.