+++ packages/stream_chat/lib/src/core/api/channel_api.dart
@@ -209,13 +209,18 @@ class ChannelApi {
     List<String> memberIds, {
     Message? message,
     bool hideHistory = false,
+    DateTime? hideHistoryBefore,
   }) async {
     final response = await _client.post(
       _getChannelUrl(channelId, channelType),
       data: {
         'add_members': memberIds,
-        'message': message,
-        'hide_history': hideHistory,
+        if (message != null) 'message': message,
+        // [hideHistoryBefore] takes precedence over [hideHistory]
+        ...switch (hideHistoryBefore?.toUtc().toIso8601String()) {
+          final hideBefore? => {'hide_history_before': hideBefore},
+          _ => {'hide_history': hideHistory},
+        },
       },
     );
     return AddMembersResponse.fromJson(response.data);
@@ -323,15 +328,32 @@ class ChannelApi {
     return EmptyResponse.fromJson(response.data);
   }
 
-  /// Marks all messages from the provided [messageId] onwards as unread
+  /// Marks the channel as unread by a given [messageId].
+  ///
+  /// All messages from the provided message onwards will be marked as unread.
   Future<EmptyResponse> markUnread(
     String channelId,
     String channelType,
-    String? messageId,
+    String messageId,
   ) async {
     final response = await _client.post(
       '${_getChannelUrl(channelId, channelType)}/unread',
       data: {'message_id': messageId},
+    );
+    return EmptyResponse.fromJson(response.data);
+  }
+
+  /// Marks the channel as unread by a given [timestamp].
+  ///
+  /// All messages after the provided timestamp will be marked as unread.
+  Future<EmptyResponse> markUnreadByTimestamp(
+    String channelId,
+    String channelType,
+    DateTime timestamp,
+  ) async {
+    final response = await _client.post(
+      '${_getChannelUrl(channelId, channelType)}/unread',
+      data: {'message_timestamp': timestamp.toUtc().toIso8601String()},
     );
     return EmptyResponse.fromJson(response.data);
   }