+++ packages/stream_chat/lib/src/core/models/channel_model.g.dart
@@ -40,6 +40,9 @@ ChannelModel _$ChannelModelFromJson(Map<String, dynamic> json) => ChannelModel(
       team: json['team'] as String?,
       cooldown: (json['cooldown'] as num?)?.toInt() ?? 0,
       messageCount: (json['message_count'] as num?)?.toInt(),
+      filterTags: (json['filter_tags'] as List<dynamic>?)
+          ?.map((e) => e as String)
+          .toList(),
     );
 
 Map<String, dynamic> _$ChannelModelToJson(ChannelModel instance) =>