+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_no.dart
@@ -549,6 +549,13 @@ class StreamChatLocalizationsNo extends GlobalStreamChatLocalizations {
       'Er du sikker på at du vil avslutte avstemningen?';
 
   @override
+  String get deletePollOptionLabel => 'Slett alternativ';
+
+  @override
+  String get deletePollOptionQuestion =>
+      'Er du sikker på at du vil slette dette alternativet?';
+
+  @override
   String get createLabel => 'Opprett';
 
   @override