+++ packages/stream_chat/CHANGELOG.md
@@ -1,3 +1,14 @@
+## 9.22.0
+
+✅ Added
+
+- Added support for `filterTags` in channels.
+- Added support for `Channel.markUnreadByTimestamp` and `Client.markChannelUnreadByTimestamp`
+  methods to mark all messages after a given timestamp as unread.
+- Added support for `hideHistoryBefore` in `Channel.addMembers` and `Client.addChannelMembers` to
+  specify a timestamp before which channel history should be hidden for newly added members. When
+  provided, it takes precedence over the `hideHistory` boolean flag.
+
 ## 10.0.0-beta.10
 
 - Included the changes from version [`9.21.0`](https://pub.dev/packages/stream_chat/changelog).