+++ packages/stream_chat_localizations/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 9.22.0
+
+- Added translations for new `deletePollOptionLabel` label.
+- Added translations for new `deletePollOptionQuestion` text.
+
 ## 10.0.0-beta.10
 
 - Included the changes from version [`9.21.0`](https://pub.dev/packages/stream_chat_localizations/changelog).